Adding host_ip config field

+++ itf/core/qemu/qemu.py
@@ -16,6 +16,9 @@ import sys
 import subprocess
 import netifaces
 import logging
+from typing import Optional
+
+from itf.core.base.target.config.ecu import Ecu
 
 logger = logging.getLogger(__name__)
 
@@ -25,42 +28,26 @@ class Qemu:
     This class shall be used to start an qemu instance based on pre-configured Qemu parameters.
     """
 
-    def __init__(
-        self,
-        path_to_image,
-        path_to_bootloader=None,
-        ram="1G",
-        cores="2",
-        cpu="Cascadelake-Server-v5",
-        host_first_network_device_ip_address="160.48.199.77",
-        host_second_network_device_ip_address="192.168.1.99",
-    ):
+    def __init__(self, config: Ecu):
         """Create a QEMU instance with the specified parameters.
-
-        :param str path_to_image: The path to the Qemu image file.
-        :param str path_to_bootloader: The path to the Qemu bootloader file.
-        :param str ram: The amount of RAM to allocate to the QEMU instance.
-        :param str cores: The number of CPU cores to allocate to the QEMU instance.
-        :param str cpu: The CPU model to emulate.
-         Default is Cascadelake-Server-v5 used to emulate modern Intel CPU features.
-         For older Ubuntu versions change that to host in case of errors.
-        :param str host_first_network_device_ip_address: The IP address of the first network device on the host.
-        :param str host_second_network_device_ip_address: The IP address of the second network device on the host.
         """
+
+        self.__config = config
         self.__qemu_path = "/usr/bin/qemu-system-x86_64"
 
         self.__first_network_device_name = "unknown"
         self.__second_network_device_name = "unknown"
         self.__first_network_adapter_mac = "52:54:11:22:33:01"
         self.__second_network_adapter_mac = "52:54:11:22:33:02"
-        self.__first_network_device_ip_address = host_first_network_device_ip_address
-        self.__second_network_device_ip_address = host_second_network_device_ip_address
 
-        self.__path_to_image = path_to_image
-        self.__path_to_bootloader = path_to_bootloader
-        self.__ram = ram
-        self.__cores = cores
-        self.__cpu = cpu
+        self.__first_network_device_ip_address = config.host_ip if config.host_ip else "160.48.199.77"
+        self.__second_network_device_ip_address: Optional[str] = config.host_diagnostic_ip
+
+        self.__path_to_image = config.qemu_image_path
+        self.__path_to_bootloader = None
+        self.__ram = config.qemu_ram_size
+        self.__cores = config.qemu_num_cores
+        self.__cpu = "Cascadelake-Server-v5"
 
         self.__check_qemu_is_installed()
         self.__find_available_kvm_support()
@@ -69,13 +56,13 @@ class Qemu:
 
         self._subprocess = None
 
-    def __enter__(self):
+    def __enter__(self) -> subprocess.Popen:
         return self.start()
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.stop()
 
-    def start(self, subprocess_params=None):
+    def start(self, subprocess_params=None) -> subprocess.Popen:
         logger.debug(self.__build_qemu_command())
         subprocess_args = {"args": shlex.split(self.__build_qemu_command())}
         if subprocess_params:
@@ -84,6 +71,9 @@ class Qemu:
         return self._subprocess
 
     def stop(self):
+        if self._subprocess is None:
+            raise RuntimeError(f"{self.__class__.__name__}.stop() called before start()")
+
         if self._subprocess.poll() is None:
             self._subprocess.terminate()
             self._subprocess.wait(2)
@@ -130,11 +120,13 @@ class Qemu:
             except KeyError:
                 pass
 
-        if "unknown" in (self.__first_network_device_name, self.__second_network_device_name):
+        if (self.__first_network_device_name == "unknown" and self.__second_network_device_name == "unknown"):
             logger.fatal("Could not find correct tap devices. Please setup network for Qemu first!")
+            logger.fatal(f"Tried to find devices with IPs: {self.__first_network_device_ip_address} {self.__second_network_device_ip_address}")
             sys.exit(-1)
+        logger.info(f"Found devices {self.__first_network_device_ip_address} {self.__second_network_device_ip_address}")
 
-    def __build_qemu_command(self):
+    def __build_qemu_command(self) -> str:
         return (
             f"{self.__qemu_path}"
             " --enable-kvm"  # Use hardware virtualization for better performance
@@ -145,18 +137,21 @@ class Qemu:
             " -nographic"  # Disable graphical display (console-only)
             " -serial mon:stdio"  # Redirect serial output to console
             " -object rng-random,filename=/dev/urandom,id=rng0"  # Provide hardware random number generation
-            f" {self.__first_network_adapter()}"
-            f" {self.__second_network_adapter()}"
+            f"{self.__first_network_adapter()}"
+            f"{self.__second_network_adapter()}"
             " -device virtio-rng-pci,rng=rng0"  # Provide hardware random number generation
         )
 
-    def __first_network_adapter(self):
+    def __first_network_adapter(self) -> str:
         return (
             f" -netdev tap,id=t1,ifname={self.__first_network_device_name},script=no,downscript=no"
             f" -device virtio-net-pci,netdev=t1,id=nic1,mac={self.__first_network_adapter_mac},guest_csum=off"
         )
 
-    def __second_network_adapter(self):
+    def __second_network_adapter(self) -> str:
+        if not self.__config.host_diagnostic_ip:
+            return ""
+
         return (
             f" -netdev tap,id=t2,ifname={self.__second_network_device_name},script=no,downscript=no"
             f" -device virtio-net-pci,netdev=t2,id=nic2,mac={self.__second_network_adapter_mac},guest_csum=off"
